Version Packages (#15)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @inqling/svelte-icons@3.0.2

### Patch Changes

-   c3c0305: Adds `SvelteIcon` type
-   c3c0305: Fixes missing simple icons export
-   cd2a103: fixes exported typings

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>