Version Packages (#26)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @inqling/svelte-icons@3.3.1

### Patch Changes

-   391fbc97: removed engine requirement

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>