Version Packages (#37)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @inqling/svelte-icons@3.4.3

### Patch Changes

-   126d0076: fixes changeset config
-   376c5850: fix svelte pinned version

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>