Version Packages (#40)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @inqling/svelte-icons@3.5.0

### Minor Changes

-   5a482d20: update major (dev) versions

### Patch Changes

-   5a482d20: upgrade minor versions

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>