Version Packages (#47)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @inqling/svelte-icons@4.1.0

### Minor Changes

-   01f55db: Add Heroicons Mini Variant
-   8a34100: Fix icon importing for icon names hat start with a number

### Patch Changes

-   0a39d80: Add 12 new Herocions
-   9f14898: linting changes
-   18600c9: Added 334 new SimpleIcons

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>