Version Packages (#49)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @inqling/svelte-icons@4.1.2

### Patch Changes

-   0a7be09: Fix export types
-   629efb4: Add or update 50 simple-icons

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>